+++ noodles-bam/src/reader/record.rs
@@ -3,7 +3,7 @@ use std::io::{self, Read};
 use byteorder::{LittleEndian, ReadBytesExt};
 use noodles_sam::{self as sam, alignment::RecordBuf};
 
-pub(crate) fn read_record<R>(
+pub(crate) fn read_record_buf<R>(
     reader: &mut R,
     header: &sam::Header,
     buf: &mut Vec<u8>,
@@ -69,7 +69,7 @@ mod tests {
     }
 
     #[test]
-    fn test_read_record() -> io::Result<()> {
+    fn test_read_record_buf() -> io::Result<()> {
         let data = [
             0x22, 0x00, 0x00, 0x00, // block_size = 34
             0xff, 0xff, 0xff, 0xff, // ref_id = -1
@@ -90,7 +90,7 @@ mod tests {
         let header = sam::Header::default();
         let mut buf = Vec::new();
         let mut record = RecordBuf::default();
-        let block_size = read_record(&mut reader, &header, &mut buf, &mut record)?;
+        let block_size = read_record_buf(&mut reader, &header, &mut buf, &mut record)?;
 
         assert_eq!(block_size, 34);
         assert_eq!(record, RecordBuf::default());