+++ noodles-bam/src/reader/record_bufs.rs
@@ -6,8 +6,8 @@ use super::Reader;
 
 /// An iterator over records of a BAM reader.
 ///
-/// This is created by calling [`Reader::records`].
-pub struct Records<'a, R>
+/// This is created by calling [`Reader::record_bufs`].
+pub struct RecordBufs<'a, R>
 where
     R: Read,
 {
@@ -16,7 +16,7 @@ where
     record: RecordBuf,
 }
 
-impl<'a, R> Records<'a, R>
+impl<'a, R> RecordBufs<'a, R>
 where
     R: Read,
 {
@@ -29,14 +29,14 @@ where
     }
 }
 
-impl<'a, R> Iterator for Records<'a, R>
+impl<'a, R> Iterator for RecordBufs<'a, R>
 where
     R: Read,
 {
     type Item = io::Result<RecordBuf>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        match self.reader.read_record(self.header, &mut self.record) {
+        match self.reader.read_record_buf(self.header, &mut self.record) {
             Ok(0) => None,
             Ok(_) => Some(Ok(self.record.clone())),
             Err(e) => Some(Err(e)),