+++ noodles-bam/src/writer.rs
@@ -235,7 +235,7 @@ mod tests {
         let mut reader = Reader::new(writer.get_ref().get_ref().as_slice());
 
         let mut record = RecordBuf::default();
-        reader.read_record(&header, &mut record)?;
+        reader.read_record_buf(&header, &mut record)?;
 
         assert!(record.name().is_none());
         assert_eq!(record.flags(), sam::record::Flags::UNMAPPED);
@@ -320,7 +320,7 @@ mod tests {
         let mut reader = Reader::new(writer.get_ref().get_ref().as_slice());
 
         let mut record = RecordBuf::default();
-        reader.read_record(&header, &mut record)?;
+        reader.read_record_buf(&header, &mut record)?;
 
         let expected = Sequence::from(b"ATCG".to_vec());
         assert_eq!(record.sequence(), &expected);
@@ -347,7 +347,7 @@ mod tests {
         let mut reader = Reader::new(writer.get_ref().get_ref().as_slice());
 
         let mut record = RecordBuf::default();
-        reader.read_record(&header, &mut record)?;
+        reader.read_record_buf(&header, &mut record)?;
 
         let expected = Sequence::from(b"ATCG".to_vec());
         assert_eq!(record.sequence(), &expected);
@@ -381,7 +381,7 @@ mod tests {
         let mut reader = Reader::new(writer.get_ref().get_ref().as_slice());
 
         let mut record = RecordBuf::default();
-        reader.read_record(&header, &mut record)?;
+        reader.read_record_buf(&header, &mut record)?;
 
         let bam_data = record.data();
         let mut fields = bam_data.iter();